+++ scripts/make_unique_peptide.py
@@ -24,7 +24,7 @@ def main():
     input_vector = l.split()[: 772]
     key = ' '.join(str(x) for x in input_vector)
     output_vector_ = [float(e) for e in l.split()[772: 782]]
-    output_vector = np.array(output_vector_).astype('float32')
+    output_vector = np.array(output_vector_).astype('float64')
     if key in feature_dict:
       feature_dict[key] += output_vector
       key_dict[key] += 1
